+++ kernel/src/device/block/partition.rs
@@ -230,7 +230,7 @@ impl MemoryMappingOps for PartitionBlockDevice {
         Err("Memory mapping not supported by partition block device")
     }
 
-    fn on_mapped(&self, _vaddr: usize, _paddr: usize, _length: usize, _offset: usize) {}
+    fn on_mapped(&self, _vaddr: usize, _paddr: u64, _length: usize, _offset: usize) {}
 
     fn on_unmapped(&self, _vaddr: usize, _length: usize) {}
