+++ kernel/src/device/gpu/connection.rs
@@ -478,7 +478,7 @@ impl MemoryMappingOps for GpuConnection {
         Err("GPU connections do not support memory mapping")
     }
 
-    fn on_mapped(&self, _vaddr: usize, _paddr: usize, _length: usize, _offset: usize) {}
+    fn on_mapped(&self, _vaddr: usize, _paddr: u64, _length: usize, _offset: usize) {}
 
     fn on_unmapped(&self, _vaddr: usize, _length: usize) {}
